Implementando outras exceções

+++ src/main/java/com/henriqueardina/dscommerce/services/ProductService.java
@@ -1,14 +1,20 @@
 package com.henriqueardina.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.henriqueardina.dscommerce.dto.ProductDTO;
 import com.henriqueardina.dscommerce.entities.Product;
 import com.henriqueardina.dscommerce.repositories.ProductRepository;
+import com.henriqueardina.dscommerce.services.exceptions.DataBaseException;
+import com.henriqueardina.dscommerce.services.exceptions.ResourceNotFoundExcepetion;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,8 +24,11 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		
+
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundExcepetion("Recurso não encontrado"));
 		return new ProductDTO(product);
+
 	}
 	
 	@Transactional(readOnly = true)
@@ -30,19 +39,23 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO insert(ProductDTO dto) {
-		Product entity = new Product();
-		copyDtoToEntuty(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);	
+			Product entity = new Product();
+			copyDtoToEntuty(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
 	}
 	
 	@Transactional
 	public ProductDTO update(Long id,ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntuty(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
-		
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntuty(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundExcepetion("Recurso não encontrado");
+		}
 	}
 
 	private void copyDtoToEntuty(ProductDTO dto, Product entity) {
@@ -53,8 +66,16 @@ public class ProductService {
 		
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundExcepetion("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DataBaseException("Falha de integridade referencial");
+	   	}
 	}
 }
